backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.0
django-environ>=0.11
psycopg2-binary>=2.9
redis>=5.0
celery>=5.3
django-redis>=5.4
channels>=4.0
channels-redis>=4.1
openai>=1.0
requests>=2.31
python-dotenv>=1.0
Pillow>=10.0
num2words>=0.5
pydub>=0.25
pytest-django>=4.5
factory-boy>=3.3
